Remove revive offer event loop
