+++ ballista/rust/scheduler/src/scheduler_server/mod.rs
@@ -20,19 +20,18 @@ use std::time::{SystemTime, UNIX_EPOCH};
 
 use ballista_core::config::TaskSchedulingPolicy;
 use ballista_core::error::Result;
-use ballista_core::event_loop::{EventAction, EventLoop};
+use ballista_core::event_loop::EventLoop;
+use ballista_core::serde::protobuf::TaskStatus;
 use ballista_core::serde::{AsExecutionPlan, BallistaCodec};
 use datafusion::execution::context::{default_session_builder, SessionState};
 use datafusion::logical_plan::LogicalPlan;
-
-use ballista_core::serde::protobuf::TaskStatus;
 use datafusion::prelude::{SessionConfig, SessionContext};
 use datafusion_proto::logical_plan::AsLogicalPlan;
 
-use crate::scheduler_server::event::{QueryStageSchedulerEvent, SchedulerServerEvent};
-use crate::scheduler_server::event_loop::SchedulerServerEventAction;
+use crate::scheduler_server::event::QueryStageSchedulerEvent;
 use crate::scheduler_server::query_stage_scheduler::QueryStageScheduler;
 use crate::state::backend::StateBackendClient;
+use crate::state::executor_manager::ExecutorReservation;
 use crate::state::SchedulerState;
 
 // include the generated protobuf source as a submodule
@@ -42,7 +41,6 @@ pub mod externalscaler {
 }
 
 pub mod event;
-mod event_loop;
 mod external_scaler;
 mod grpc;
 mod query_stage_scheduler;
@@ -55,7 +53,6 @@ pub struct SchedulerServer<T: 'static + AsLogicalPlan, U: 'static + AsExecutionP
     pub(crate) state: Arc<SchedulerState<T, U>>,
     pub start_time: u128,
     policy: TaskSchedulingPolicy,
-    event_loop: Option<EventLoop<SchedulerServerEvent>>,
     pub(crate) query_stage_event_loop: EventLoop<QueryStageSchedulerEvent>,
 }
 
@@ -103,30 +100,16 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan> SchedulerServer<T
             scheduler_name.clone(),
         ));
 
-        let event_action: Option<Arc<dyn EventAction<SchedulerServerEvent>>> =
-            if matches!(policy, TaskSchedulingPolicy::PushStaged) {
-                Some(Arc::new(SchedulerServerEventAction::new(state.clone())))
-            } else {
-                None
-            };
-        SchedulerServer::new_with_event_action(scheduler_name, state, event_action)
+        SchedulerServer::new_with_state(scheduler_name, policy, state)
     }
 
-    fn new_with_event_action(
+    pub(crate) fn new_with_state(
         scheduler_name: String,
+        policy: TaskSchedulingPolicy,
         state: Arc<SchedulerState<T, U>>,
-        event_action: Option<Arc<dyn EventAction<SchedulerServerEvent>>>,
     ) -> Self {
-        let event_loop = event_action.map(|event_action| {
-            EventLoop::new("scheduler".to_owned(), 10000, event_action)
-        });
-        let policy = if event_loop.is_some() {
-            TaskSchedulingPolicy::PushStaged
-        } else {
-            TaskSchedulingPolicy::PullStaged
-        };
         let query_stage_scheduler =
-            Arc::new(QueryStageScheduler::new(state.clone(), None));
+            Arc::new(QueryStageScheduler::new(state.clone(), policy));
         let query_stage_event_loop =
             EventLoop::new("query_stage".to_owned(), 10000, query_stage_scheduler);
         Self {
@@ -137,35 +120,13 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan> SchedulerServer<T
                 .unwrap()
                 .as_millis(),
             policy,
-            event_loop,
             query_stage_event_loop,
         }
     }
 
     pub async fn init(&mut self) -> Result<()> {
-        {
-            // initialize state
-            self.state.init().await?;
-        }
-
-        {
-            if let Some(event_loop) = self.event_loop.as_mut() {
-                event_loop.start()?;
-
-                let query_stage_scheduler = Arc::new(QueryStageScheduler::new(
-                    self.state.clone(),
-                    Some(event_loop.get_sender()?),
-                ));
-                let query_stage_event_loop = EventLoop::new(
-                    self.query_stage_event_loop.name.clone(),
-                    self.query_stage_event_loop.buffer_size,
-                    query_stage_scheduler,
-                );
-                self.query_stage_event_loop = query_stage_event_loop;
-            }
-
-            self.query_stage_event_loop.start()?;
-        }
+        self.state.init().await?;
+        self.query_stage_event_loop.start()?;
 
         Ok(())
     }
@@ -201,6 +162,16 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan> SchedulerServer<T
             ))
             .await
     }
+
+    pub(crate) async fn offer_reservation(
+        &self,
+        reservations: Vec<ExecutorReservation>,
+    ) -> Result<()> {
+        self.query_stage_event_loop
+            .get_sender()?
+            .post_event(QueryStageSchedulerEvent::ReservationOffering(reservations))
+            .await
+    }
 }
 
 #[cfg(all(test, feature = "sled"))]
@@ -218,8 +189,7 @@ mod test {
     use ballista_core::config::{
         BallistaConfig, TaskSchedulingPolicy, BALLISTA_DEFAULT_SHUFFLE_PARTITIONS,
     };
-    use ballista_core::error::{BallistaError, Result};
-    use ballista_core::event_loop::EventAction;
+    use ballista_core::error::Result;
 
     use ballista_core::serde::protobuf::{
         job_status, task_status, CompletedTask, FailedTask, JobStatus, PartitionId,
@@ -230,15 +200,12 @@ mod test {
     };
     use ballista_core::serde::BallistaCodec;
 
-    use crate::scheduler_server::event::SchedulerServerEvent;
     use crate::scheduler_server::SchedulerServer;
     use crate::state::backend::standalone::StandaloneClient;
 
     use crate::state::executor_manager::ExecutorReservation;
     use crate::state::SchedulerState;
-    use crate::test_utils::{
-        await_condition, ExplodingTableProvider, SchedulerEventObserver,
-    };
+    use crate::test_utils::{await_condition, ExplodingTableProvider};
 
     #[tokio::test]
     async fn test_pull_scheduling() -> Result<()> {
@@ -264,33 +231,12 @@ mod test {
             .create_session(&config)
             .await?;
 
-        let plan = async {
-            let optimized_plan = ctx.optimize(&plan).map_err(|e| {
-                BallistaError::General(format!(
-                    "Could not create optimized logical plan: {}",
-                    e
-                ))
-            })?;
-
-            ctx.create_physical_plan(&optimized_plan)
-                .await
-                .map_err(|e| {
-                    BallistaError::General(format!(
-                        "Could not create physical plan: {}",
-                        e
-                    ))
-                })
-        }
-        .await?;
-
         let job_id = "job";
-        let session_id = ctx.session_id();
 
         // Submit job
         scheduler
             .state
-            .task_manager
-            .submit_job(job_id, &session_id, plan)
+            .submit_job(job_id, ctx, &plan)
             .await
             .expect("submitting plan");
 
@@ -373,13 +319,7 @@ mod test {
         let plan = test_plan();
         let task_slots = 4;
 
-        let (sender, mut event_receiver) =
-            tokio::sync::mpsc::channel::<SchedulerServerEvent>(1000);
-        let (error_sender, _) = tokio::sync::mpsc::channel::<BallistaError>(1000);
-
-        let event_action = SchedulerEventObserver::new(sender, error_sender);
-
-        let scheduler = test_scheduler_with_event_action(Arc::new(event_action)).await?;
+        let scheduler = test_push_staged_scheduler().await?;
 
         let executors = test_executors(task_slots);
         for (executor_metadata, executor_data) in executors {
@@ -400,13 +340,39 @@ mod test {
 
         let job_id = "job";
 
-        // Send JobQueued event to kick off the event loop
-        scheduler.submit_job(job_id, ctx, &plan).await?;
+        scheduler.state.submit_job(job_id, ctx, &plan).await?;
 
         // Complete tasks that are offered through scheduler events
-        while let Some(SchedulerServerEvent::Offer(reservations)) =
-            event_receiver.recv().await
-        {
+        loop {
+            // Check condition
+            let available_tasks = {
+                let graph = scheduler
+                    .state
+                    .task_manager
+                    .get_active_execution_graph(job_id)
+                    .await
+                    .unwrap();
+                let graph = graph.read().await;
+                if graph.complete() {
+                    break;
+                }
+                graph.available_tasks()
+            };
+
+            if available_tasks == 0 {
+                tokio::time::sleep(Duration::from_millis(5)).await;
+                continue;
+            }
+
+            let reservations: Vec<ExecutorReservation> = scheduler
+                .state
+                .executor_manager
+                .reserve_slots(available_tasks as u32)
+                .await?
+                .into_iter()
+                .map(|res| res.assign(job_id.to_owned()))
+                .collect();
+
             let free_list = match scheduler
                 .state
                 .task_manager
@@ -414,11 +380,6 @@ mod test {
                 .await
             {
                 Ok((assignments, mut unassigned_reservations, _)) => {
-                    // Break when we are no longer assigning tasks
-                    if unassigned_reservations.len() == reservations.len() {
-                        break;
-                    }
-
                     for (executor_id, task) in assignments.into_iter() {
                         match scheduler
                             .state
@@ -504,13 +465,7 @@ mod test {
         let plan = test_plan();
         let task_slots = 4;
 
-        let (sender, mut event_receiver) =
-            tokio::sync::mpsc::channel::<SchedulerServerEvent>(1000);
-        let (error_sender, _) = tokio::sync::mpsc::channel::<BallistaError>(1000);
-
-        let event_action = SchedulerEventObserver::new(sender, error_sender);
-
-        let scheduler = test_scheduler_with_event_action(Arc::new(event_action)).await?;
+        let scheduler = test_push_staged_scheduler().await?;
 
         let executors = test_executors(task_slots);
         for (executor_metadata, executor_data) in executors {
@@ -531,87 +486,92 @@ mod test {
 
         let job_id = "job";
 
-        // Send JobQueued event to kick off the event loop
-        scheduler.submit_job(job_id, ctx, &plan).await?;
+        scheduler.state.submit_job(job_id, ctx, &plan).await?;
+
+        let available_tasks = scheduler
+            .state
+            .task_manager
+            .get_available_task_count(job_id)
+            .await?;
+
+        let reservations: Vec<ExecutorReservation> = scheduler
+            .state
+            .executor_manager
+            .reserve_slots(available_tasks as u32)
+            .await?
+            .into_iter()
+            .map(|res| res.assign(job_id.to_owned()))
+            .collect();
 
         // Complete tasks that are offered through scheduler events
-        if let Some(SchedulerServerEvent::Offer(reservations)) =
-            event_receiver.recv().await
+        let free_list = match scheduler
+            .state
+            .task_manager
+            .fill_reservations(&reservations)
+            .await
         {
-            let free_list = match scheduler
-                .state
-                .task_manager
-                .fill_reservations(&reservations)
-                .await
-            {
-                Ok((assignments, mut unassigned_reservations, _)) => {
-                    for (executor_id, task) in assignments.into_iter() {
-                        match scheduler
-                            .state
-                            .executor_manager
-                            .get_executor_metadata(&executor_id)
-                            .await
-                        {
-                            Ok(executor) => {
-                                let mut partitions: Vec<ShuffleWritePartition> = vec![];
+            Ok((assignments, mut unassigned_reservations, _)) => {
+                for (executor_id, task) in assignments.into_iter() {
+                    match scheduler
+                        .state
+                        .executor_manager
+                        .get_executor_metadata(&executor_id)
+                        .await
+                    {
+                        Ok(executor) => {
+                            let mut partitions: Vec<ShuffleWritePartition> = vec![];
 
-                                let num_partitions = task
-                                    .output_partitioning
-                                    .map(|p| p.partition_count())
-                                    .unwrap_or(1);
+                            let num_partitions = task
+                                .output_partitioning
+                                .map(|p| p.partition_count())
+                                .unwrap_or(1);
 
-                                for partition_id in 0..num_partitions {
-                                    partitions.push(ShuffleWritePartition {
-                                        partition_id: partition_id as u64,
-                                        path: "some/path".to_string(),
-                                        num_batches: 1,
-                                        num_rows: 1,
-                                        num_bytes: 1,
-                                    })
-                                }
-
-                                // Complete the task
-                                let task_status = TaskStatus {
-                                    status: Some(task_status::Status::Failed(
-                                        FailedTask {
-                                            error: "".to_string(),
-                                        },
-                                    )),
-                                    metrics: vec![],
-                                    task_id: Some(PartitionId {
-                                        job_id: job_id.to_owned(),
-                                        stage_id: task.partition.stage_id as u32,
-                                        partition_id: task.partition.partition_id as u32,
-                                    }),
-                                };
-
-                                scheduler
-                                    .state
-                                    .update_task_statuses(&executor.id, vec![task_status])
-                                    .await?;
+                            for partition_id in 0..num_partitions {
+                                partitions.push(ShuffleWritePartition {
+                                    partition_id: partition_id as u64,
+                                    path: "some/path".to_string(),
+                                    num_batches: 1,
+                                    num_rows: 1,
+                                    num_bytes: 1,
+                                })
                             }
-                            Err(_e) => {
-                                unassigned_reservations.push(
-                                    ExecutorReservation::new_free(executor_id.clone()),
-                                );
-                            }
+
+                            // Complete the task
+                            let task_status = TaskStatus {
+                                status: Some(task_status::Status::Failed(FailedTask {
+                                    error: "".to_string(),
+                                })),
+                                metrics: vec![],
+                                task_id: Some(PartitionId {
+                                    job_id: job_id.to_owned(),
+                                    stage_id: task.partition.stage_id as u32,
+                                    partition_id: task.partition.partition_id as u32,
+                                }),
+                            };
+
+                            scheduler
+                                .state
+                                .update_task_statuses(&executor.id, vec![task_status])
+                                .await?;
+                        }
+                        Err(_e) => {
+                            unassigned_reservations
+                                .push(ExecutorReservation::new_free(executor_id.clone()));
                         }
                     }
-                    unassigned_reservations
                 }
-                Err(_e) => reservations,
-            };
-
-            // If any reserved slots remain, return them to the pool
-            if !free_list.is_empty() {
-                scheduler
-                    .state
-                    .executor_manager
-                    .cancel_reservations(free_list)
-                    .await?;
+                unassigned_reservations
             }
-        } else {
-            panic!("No reservations offered");
+            Err(_e) => reservations,
+        };
+
+        // If any reserved slots remain, return them to the pool
+        if !free_list.is_empty() {
+            scheduler
+                .state
+                .executor_manager
+                .cancel_reservations(free_list)
+                .await?;
         }
 
         let status = scheduler.state.task_manager.get_job_status(job_id).await?;
@@ -635,13 +595,7 @@ mod test {
     async fn test_planning_failure() -> Result<()> {
         let task_slots = 4;
 
-        let (sender, _event_receiver) =
-            tokio::sync::mpsc::channel::<SchedulerServerEvent>(1000);
-        let (error_sender, _) = tokio::sync::mpsc::channel::<BallistaError>(1000);
-
-        let event_action = SchedulerEventObserver::new(sender, error_sender);
-
-        let scheduler = test_scheduler_with_event_action(Arc::new(event_action)).await?;
+        let scheduler = test_push_staged_scheduler().await?;
 
         let config = test_session(task_slots);
 
@@ -657,7 +611,6 @@ mod test {
 
         let job_id = "job";
 
-        // Send JobQueued event to kick off the event loop
         // This should fail when we try and create the physical plan
         scheduler.submit_job(job_id, ctx, &plan).await?;
 
@@ -699,8 +652,7 @@ mod test {
         Ok(scheduler)
     }
 
-    async fn test_scheduler_with_event_action(
-        event_action: Arc<dyn EventAction<SchedulerServerEvent>>,
+    async fn test_push_staged_scheduler(
     ) -> Result<SchedulerServer<LogicalPlanNode, PhysicalPlanNode>> {
         let state_storage = Arc::new(StandaloneClient::try_new_temporary()?);
         let state = Arc::new(SchedulerState::new_with_default_scheduler_name(
@@ -709,10 +661,10 @@ mod test {
             BallistaCodec::default(),
         ));
         let mut scheduler: SchedulerServer<LogicalPlanNode, PhysicalPlanNode> =
-            SchedulerServer::new_with_event_action(
+            SchedulerServer::new_with_state(
                 "localhost:50050".to_owned(),
+                TaskSchedulingPolicy::PushStaged,
                 state,
-                Some(event_action),
             );
         scheduler.init().await?;
