+++ ballista/rust/scheduler/src/scheduler_server/query_stage_scheduler.rs
@@ -16,21 +16,19 @@
 // under the License.
 
 use std::sync::Arc;
-use std::time::Instant;
 
 use async_trait::async_trait;
-use datafusion::logical_plan::LogicalPlan;
-use datafusion::prelude::SessionContext;
 use log::{debug, error, info};
 
 use ballista_core::error::{BallistaError, Result};
 use ballista_core::event_loop::{EventAction, EventSender};
 
+use ballista_core::config::TaskSchedulingPolicy;
 use ballista_core::serde::AsExecutionPlan;
 use datafusion_proto::logical_plan::AsLogicalPlan;
 use tokio::sync::mpsc;
 
-use crate::scheduler_server::event::{QueryStageSchedulerEvent, SchedulerServerEvent};
+use crate::scheduler_server::event::QueryStageSchedulerEvent;
 
 use crate::state::executor_manager::ExecutorReservation;
 use crate::state::SchedulerState;
@@ -40,18 +38,15 @@ pub(crate) struct QueryStageScheduler<
     U: 'static + AsExecutionPlan,
 > {
     state: Arc<SchedulerState<T, U>>,
-    event_sender: Option<EventSender<SchedulerServerEvent>>,
+    policy: TaskSchedulingPolicy,
 }
 
 impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan> QueryStageScheduler<T, U> {
     pub(crate) fn new(
         state: Arc<SchedulerState<T, U>>,
-        event_sender: Option<EventSender<SchedulerServerEvent>>,
+        policy: TaskSchedulingPolicy,
     ) -> Self {
-        Self {
-            state,
-            event_sender,
-        }
+        Self { state, policy }
     }
 }
 
@@ -73,6 +68,7 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
         tx_event: &mpsc::Sender<QueryStageSchedulerEvent>,
         _rx_event: &mpsc::Receiver<QueryStageSchedulerEvent>,
     ) -> Result<()> {
+        let tx_event = EventSender::new(tx_event.clone());
         match event {
             QueryStageSchedulerEvent::JobQueued {
                 job_id,
@@ -81,11 +77,9 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
             } => {
                 info!("Job {} queued", job_id);
                 let state = self.state.clone();
-                let tx_event = tx_event.clone();
                 tokio::spawn(async move {
                     let event = if let Err(e) =
-                        submit_job(state.clone(), job_id.clone(), session_ctx, &plan)
-                            .await
+                        state.submit_job(&job_id, session_ctx, &plan).await
                     {
                         let msg = format!("Error planning job {}: {:?}", job_id, e);
                         error!("{}", &msg);
@@ -94,17 +88,15 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
                         QueryStageSchedulerEvent::JobSubmitted(job_id)
                     };
                     tx_event
-                        .send(event)
+                        .post_event(event)
                         .await
-                        .map_err(|e| {
-                            error!("Fail to send event due to {}", e);
-                        })
+                        .map_err(|e| error!("Fail to send event due to {}", e))
                         .unwrap();
                 });
             }
             QueryStageSchedulerEvent::JobSubmitted(job_id) => {
                 info!("Job {} submitted", job_id);
-                if let Some(sender) = &self.event_sender {
+                if matches!(self.policy, TaskSchedulingPolicy::PushStaged) {
                     let available_tasks = self
                         .state
                         .task_manager
@@ -126,16 +118,11 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
                         job_id
                     );
 
-                    if let Err(e) = sender
-                        .post_event(SchedulerServerEvent::Offer(reservations.clone()))
-                        .await
-                    {
-                        error!("Error posting offer: {:?}", e);
-                        self.state
-                            .executor_manager
-                            .cancel_reservations(reservations)
-                            .await?;
-                    }
+                    tx_event
+                        .post_event(QueryStageSchedulerEvent::ReservationOffering(
+                            reservations,
+                        ))
+                        .await?;
                 }
             }
             QueryStageSchedulerEvent::JobPlanningFailed(job_id, fail_message) => {
@@ -165,20 +152,16 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
                     .await
                 {
                     Ok((stage_events, offers)) => {
-                        if let Some(event_sender) = self.event_sender.as_ref() {
-                            event_sender
-                                .post_event(SchedulerServerEvent::Offer(offers))
+                        if matches!(self.policy, TaskSchedulingPolicy::PushStaged) {
+                            tx_event
+                                .post_event(
+                                    QueryStageSchedulerEvent::ReservationOffering(offers),
+                                )
                                 .await?;
                         }
 
                         for stage_event in stage_events {
-                            tx_event
-                                .send(stage_event)
-                                .await
-                                .map_err(|e| {
-                                    error!("Fail to send event due to {}", e);
-                                })
-                                .unwrap();
+                            tx_event.post_event(stage_event).await?;
                         }
                     }
                     Err(e) => {
@@ -190,6 +173,16 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
                     }
                 }
             }
+            QueryStageSchedulerEvent::ReservationOffering(reservations) => {
+                let reservations = self.state.offer_reservation(reservations).await?;
+                if !reservations.is_empty() {
+                    tx_event
+                        .post_event(QueryStageSchedulerEvent::ReservationOffering(
+                            reservations,
+                        ))
+                        .await?;
+                }
+            }
         }
 
         Ok(())
@@ -198,29 +191,4 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>
     fn on_error(&self, error: BallistaError) {
         error!("Error received by QueryStageScheduler: {:?}", error);
     }
-}
-
-async fn submit_job<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan>(
-    state: Arc<SchedulerState<T, U>>,
-    job_id: String,
-    session_ctx: Arc<SessionContext>,
-    plan: &LogicalPlan,
-) -> Result<()> {
-    let start = Instant::now();
-    let optimized_plan = session_ctx.optimize(plan)?;
-
-    debug!("Calculated optimized plan: {:?}", optimized_plan);
-
-    let plan = session_ctx.create_physical_plan(&optimized_plan).await?;
-
-    state
-        .task_manager
-        .submit_job(&job_id, &session_ctx.session_id(), plan.clone())
-        .await?;
-
-    let elapsed = start.elapsed();
-
-    info!("Planned job {} in {:?}", job_id, elapsed);
-
-    Ok(())
 }
