+++ ballista/rust/scheduler/src/test_utils.rs
@@ -15,16 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use ballista_core::error::{BallistaError, Result};
+use ballista_core::error::Result;
 use std::any::Any;
 use std::future::Future;
 use std::sync::Arc;
 use std::time::Duration;
 
-use crate::scheduler_server::event::SchedulerServerEvent;
-
 use async_trait::async_trait;
-use ballista_core::event_loop::EventAction;
 
 use datafusion::arrow::datatypes::{DataType, Field, Schema, SchemaRef};
 use datafusion::common::DataFusionError;
@@ -33,49 +30,10 @@ use datafusion::execution::context::{SessionConfig, SessionContext, SessionState
 use datafusion::logical_expr::Expr;
 use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::CsvReadOptions;
-use tokio::sync::mpsc::{Receiver, Sender};
 
 pub const TPCH_TABLES: &[&str] = &[
     "part", "supplier", "partsupp", "customer", "orders", "lineitem", "nation", "region",
 ];
-
-/// Test utility that allows observing scheduler events.
-pub struct SchedulerEventObserver {
-    sender: Sender<SchedulerServerEvent>,
-    errors: Sender<BallistaError>,
-}
-
-impl SchedulerEventObserver {
-    pub fn new(
-        sender: Sender<SchedulerServerEvent>,
-        errors: Sender<BallistaError>,
-    ) -> Self {
-        Self { sender, errors }
-    }
-}
-
-#[async_trait]
-impl EventAction<SchedulerServerEvent> for SchedulerEventObserver {
-    fn on_start(&self) {}
-
-    fn on_stop(&self) {}
-
-    async fn on_receive(
-        &self,
-        event: SchedulerServerEvent,
-        _tx_event: &Sender<SchedulerServerEvent>,
-        _rx_event: &Receiver<SchedulerServerEvent>,
-    ) -> Result<()> {
-        self.sender.send(event).await.unwrap();
-
-        Ok(())
-    }
-
-    fn on_error(&self, error: BallistaError) {
-        let errors = self.errors.clone();
-        tokio::task::spawn(async move { errors.send(error).await.unwrap() });
-    }
-}
 
 /// Sometimes we need to construct logical plans that will produce errors
 /// when we try and create physical plan. A scan using `ExplodingTableProvider`
